Added transitions to icons on FE. Started work on achievements.

+++ resources/js/Pages/Word.vue
@@ -2,31 +2,54 @@
 import {Head} from '@inertiajs/vue3'
 import {
     PhArrowsCounterClockwise,
-    PhCaretDown,
-    PhMinus,
-    PhPlus,
+    PhCaretDown, PhHeart,
     PhThumbsDown,
     PhThumbsUp, PhTrash,
-    PhX
 } from "@phosphor-icons/vue";
-import {computed, reactive, ref} from "vue";
+import {computed, onMounted, onUnmounted, reactive, ref} from "vue";
 
-const props = defineProps(['word', 'parsedDefinitions', 'categories', 'favorited', 'favorites']);
+const props = defineProps(['word', 'parsedDefinitions', 'categories', 'favorited', 'favorites', 'vote']);
 
+// keydown listen for spacebar
+const handleKeyDown = (event) => {
+    if (event.code === "Space") {
+        event.preventDefault();
+        refreshWord();
+    }
+};
+
+onMounted(() => {
+    window.addEventListener("keydown", handleKeyDown);
+    // set vote
+    if (props.vote !== null) {
+        if (props.vote === true) {
+            voteState.upvoted = true;
+            voteState.downvoted = false;
+        } else {
+            voteState.upvoted = false;
+            voteState.downvoted = true;
+        }
+    }
+});
+
+onUnmounted(() => {
+    window.removeEventListener("keydown", handleKeyDown);
+});
+
+// Reactive variables
 const word = ref(props.word);
 const parsedDefinitions = ref(props.parsedDefinitions);
 const categories = ref(props.categories);
 const favorites = ref(props.favorites);
 const favorited = ref(props.favorited);
-
-
+const isRotating = ref(false);
 const vote_count = computed(() => props.word.vote_count);
 const voteState = reactive({
     upvoted: false,
     downvoted: false
 });
 
-const refresh = async () => {
+const refreshWord = async () => {
     // send in the current id to avoid it on refresh...
     // todo make an achievement to get the same word twice in a row perhaps!
     try {
@@ -38,33 +61,55 @@ const refresh = async () => {
         favorites.value = response.data.favorites;
         favorited.value = response.data.favorited;
 
+        // update vote state
+        if (response.data.vote !== null) {
+            if (response.data.vote === true) {
+                voteState.upvoted = true;
+                voteState.downvoted = false;
+            } else {
+                voteState.upvoted = false;
+                voteState.downvoted = true;
+            }
+        } else {
+            voteState.upvoted = false;
+            voteState.downvoted = false;
+        }
+
     } catch (error) {
         console.error('Error refreshing word:', error);
         alert('Failed to refresh word. Please try again.');
     }
 
 };
-const addFavorite = async (wordId) => {
+
+// rotate the refresh button
+const startRotationAndRefreshWord = () => {
+    isRotating.value = true;
+    refreshWord().finally(() => {
+        // Stop rotating after a delay (animation duration)
+        setTimeout(() => {
+            isRotating.value = false;
+        }, 1000); // Match this duration to your CSS animation
+    });
+};
+
+const toggleFavorite = async (wordId) => {
     try {
-        const response = await axios.post(`/favorites/${wordId}/add`);
-        favorites.value = response.data.favorites;
-        favorited.value = true;
+        if (favorited.value) {
+            const response = await axios.delete(`/favorites/${wordId}/remove`);
+            favorites.value = response.data.favorites;
+            favorited.value = false;
+        } else {
+            const response = await axios.post(`/favorites/${wordId}/add`);
+            favorites.value = response.data.favorites;
+            favorited.value = true;
+        }
     } catch (error) {
-        console.error('Error adding favorite:', error);
-        alert('Failed to add favorite. Please try again.');
+        console.error('Error toggling favorite:', error);
+        alert('Failed to toggle favorite. Please try again.');
     }
 };
 
-const removeFavorite = async (wordId) => {
-    try {
-        const response = await axios.delete(`/favorites/${wordId}/remove`);
-        favorites.value = response.data.favorites;
-        favorited.value = false;
-    } catch (error) {
-        console.error('Error removing favorite:', error);
-        alert('Failed to remove favorite. Please try again.');
-    }
-};
 
 const clearFavorites = async () => {
     try {
@@ -108,7 +153,7 @@ const vote = async (wordId, voteValue) => {
 <template>
     <Head :title="`quasi | ${word.name}`"/>
 
-    <div class="flex flex-col min-h-dvh selection:bg-amber-200">
+    <div class="flex flex-col min-h-dvh selection:bg-amber-200 bg-neutral-50">
         <div class="container mx-auto max-w-screen-lg px-16 lg:px-1">
             <header class="flex flex-row items-center justify-between border-b-2 border-b-amber-400 mt-8">
                 <h1 class="text-4xl font-bold italic mb-4">
@@ -118,59 +163,91 @@ const vote = async (wordId, voteValue) => {
 
                     <!-- Upvote Button -->
                     <button v-if="vote" @click="vote(word.id, 1)"
-                            class="outline-amber-400 relative group hover:bg-amber-50 p-1 rounded-xl">
-                        <ph-thumbs-up v-if="voteState.upvoted" weight="fill" class="text-2xl text-primary-600"/>
-                        <ph-thumbs-up v-else weight="duotone" class="text-2xl text-primary-600"/>
+                            class="outline-amber-400 relative group hover:bg-amber-50 max-w-8 max-h-8 p-1 rounded-xl">
+                        <Transition name="icon" mode="out-in">
+                            <ph-thumbs-up
+                                v-if="voteState.upvoted"
+                                key="upvoted"
+                                weight="fill"
+                                :size="24"
+                            />
+                            <ph-thumbs-up
+                                v-else
+                                key="not-upvoted"
+                                weight="duotone"
+                                :size="24"
+                            />
+                        </Transition>
                         <span
                             class="absolute z-50 opacity-0 px-3 py-2 text-sm font-medium text-white bg-gray-800 rounded-lg shadow-lg group-hover:block group-hover:opacity-100 top-full left-1/2 transform -translate-x-1/2 mt-1 transition-opacity">
-                        Seen this before
+                        Seen before (outside of quasi)
                     </span>
                     </button>
 
                     <!-- Downvote Button -->
                     <button v-if="vote" @click="vote(word.id, 0)"
                             class="outline-amber-400 relative group hover:bg-amber-50 p-1 rounded-xl">
-                        <ph-thumbs-down v-if="voteState.downvoted" weight="fill" class="text-2xl text-primary-600"/>
-                        <ph-thumbs-down v-else weight="duotone" class="text-2xl text-primary-600"/>
+                        <Transition name="icon" mode="out-in">
+                            <ph-thumbs-down
+                                v-if="voteState.downvoted"
+                                key="downvoted"
+                                weight="fill"
+                                :size="24"
+                            />
+                            <ph-thumbs-down
+                                v-else
+                                key="not-downvoted"
+                                weight="duotone"
+                                :size="24"
+                            />
+                        </Transition>
                         <span
                             class="absolute z-50 opacity-0 px-3 py-2 text-sm font-medium text-white bg-gray-800 rounded-lg shadow-lg group-hover:block group-hover:opacity-100 top-full left-1/2 transform -translate-x-1/2 mt-1 transition-opacity">
-                        Never seen before
+                        Never seen (outside of quasi)
                     </span>
                     </button>
 
-                    <button v-if="!favorited" @click="addFavorite(word.id)"
+                    <button @click="toggleFavorite(word.id)"
                             class="outline-amber-400 relative group hover:bg-amber-50 p-1 rounded-xl">
-                        <ph-plus :size="24"/>
+                        <Transition name="icon" mode="out-in">
+                            <ph-heart
+                                v-if="!favorited"
+                                key="not-favorited"
+                                weight="duotone"
+                                :size="24"
+                            />
+                            <ph-heart
+                                v-else
+                                key="favorited"
+                                weight="fill"
+                                class="text-red-400"
+                                :size="24"
+                            />
+                        </Transition>
                         <span
-                            class="absolute z-50 opacity-0 px-3 py-2 text-sm font-medium text-white bg-gray-800 rounded-lg shadow-lg group-hover:block group-hover:opacity-100 top-full left-1/2 transform -translate-x-1/2 mt-1 transition-opacity"
-                        >
-                        Add to favorites
-                    </span>
+                            class="absolute z-50 opacity-0 px-3 py-2 text-sm font-medium text-white bg-gray-800 rounded-lg shadow-lg group-hover:block group-hover:opacity-100 top-full left-1/2 transform -translate-x-1/2 mt-1 transition-opacity">
+                                {{ favorited ? 'Remove from favorites' : 'Add to favorites' }}
+                            </span>
                     </button>
-                    <button v-else @click="removeFavorite(word.id)"
+
+                    <button @click="startRotationAndRefreshWord"
                             class="outline-amber-400 relative group hover:bg-amber-50 p-1 rounded-xl">
-                        <ph-minus :size="24"/>
+                        <ph-arrows-counter-clockwise
+                            :class="{ rotating: isRotating }"
+                            :size="24"
+                        />
                         <span
-                            class="absolute z-50 opacity-0 px-3 py-2 text-sm font-medium text-white bg-gray-800 rounded-lg shadow-lg group-hover:block group-hover:opacity-100 top-full left-1/2 transform -translate-x-1/2 mt-1 transition-opacity"
-                        >
-                        Remove from favorites
-                    </span>
-                    </button>
-                    <button @click="refresh"
-                            class="outline-amber-400 relative group hover:bg-amber-50 p-1 rounded-xl">
-                        <ph-arrows-counter-clockwise :size="24"/>
-                        <span
-                            class="absolute z-50 opacity-0 px-3 py-2 text-sm font-medium text-white bg-gray-800 rounded-lg shadow-lg group-hover:block group-hover:opacity-100 top-full left-1/2 transform -translate-x-1/2 mt-1 transition-opacity"
-                        >
-                        Refresh word
-                    </span>
+                            class="absolute z-50 opacity-0 px-3 py-2 text-sm font-medium text-white bg-gray-800 rounded-lg shadow-lg group-hover:block group-hover:opacity-100 top-full left-1/2 transform -translate-x-1/2 mt-1 transition-opacity">
+                                Refresh word
+                            </span>
                     </button>
                 </div>
             </header>
             <main class="mb-8">
                 <article class="my-8 grid grid-cols-1 place-items-baseline justify-items-center">
                     <!-- Word -->
-                    <section class="flex flex-col justify-center text-center sticky top-0 bg-white w-full z-10 p-4">
+                    <section
+                        class="flex flex-col justify-center text-center sticky top-0 bg-neutral-50 w-full z-10 p-4">
                         <h2 class="text-3xl font-bold">{{ word.name }}
                         </h2>
                         <small class="text-lg text-gray-500">{{ word.ipa_pronunciation }}</small>
@@ -239,7 +316,7 @@ const vote = async (wordId, voteValue) => {
             </main>
         </div>
         <!-- TODO Animate this -->
-        <footer v-if="favorites.length > 0" class="my-8 sticky bottom-0 bg-white w-full z-10 px-16 py-8">
+        <footer v-if="favorites.length > 0" class="my-8 sticky bottom-0 bg-neutral-50 w-full z-10 px-16 py-8">
             <!-- Favorites list -->
             <section class="border-b-2 border-amber-400">
                 <header class="flex flex-row justify-between itesm-center">
@@ -262,8 +339,9 @@ const vote = async (wordId, voteValue) => {
                 <div class="overflow-x-auto scrollbar-hidden">
                     <ul class="flex flex-row flex-nowrap gap-8">
                         <li v-for="favorite in favorites" :key="favorite.id"
-                            class="flex-shrink-0 w-40 bg-white z-10 p-4 text-center">
-                            <p class="text-lg font-bold" :class="{ 'bg-amber-400 rounded-lg': favorite.id === word.id}">{{ favorite.name }}</p>
+                            class="flex-shrink-0 w-40 bg-neutral-50 z-10 p-4 text-center">
+                            <p class="text-lg font-bold" :class="{ 'bg-amber-400 rounded-lg': favorite.id === word.id}">
+                                {{ favorite.name }}</p>
                             <small class="text-sm text-gray-500">Frequency: {{ favorite.frequency }}</small>
                         </li>
                     </ul>
@@ -274,4 +352,43 @@ const vote = async (wordId, voteValue) => {
 </template>
 
 <style scoped>
+.icon-enter-active, .icon-leave-active {
+    transition: all 0.222s cubic-bezier(1, 0.5, 0.8, 1);
+}
+
+.icon-enter-from {
+    transform: rotate(-8deg) translateY(-7px);
+    opacity: 0;
+}
+
+.icon-enter-to {
+    transform: rotate(0deg);
+    opacity: 1;
+}
+
+.icon-leave-from {
+    transform: rotate(0deg);
+    opacity: 1;
+}
+
+.icon-leave-to {
+    transform: rotate(-8deg) translateY(-4px);
+    opacity: 0;
+}
+
+
+.rotating {
+    animation: rotate 550ms linear;
+}
+
+@keyframes rotate {
+    from {
+        transform: rotate(0deg);
+    }
+    to {
+        transform: rotate(360deg);
+    }
+}
+
+
 </style>
